Refactor application to work in heroku New heroku procfile and requirements Unzipped ml-model made with 25000 rows Change app starting point to root run.py file Add null check to features, prediction is not done if weather forecast gets nulls
