+++ .gitignore
@@ -1,2 +1,3 @@
 /data
 __pycache__/
+.idea
