+++ .slugignore
@@ -1,0 +1,4 @@
+utils
+notebooks
+data
+.idea
