--- app/README.md
@@ -1,5 +1,0 @@
-# Train predictor app
-
-First run model.py to get `modeltrain.py` file to test app if not exists.
-
-Then run locally with `python app.py` and go to `localhost:5000/home`.
