+++ application/README.md
@@ -1,0 +1,5 @@
+# Train predictor app
+
+If the application does not yet have a ml-model, first run from root folder `python application/model.py` to create the machine learning model. For this you will need folder and file `data/trains_and_weather.csv` with the train and weather data.
+
+Then run application from root folder with command `python run.py` and go to `localhost:5000/home`.
