+++ application/templates/statistics.html
@@ -1,0 +1,65 @@
+<!DOCTYPE html>
+<html >
+<head>
+<meta charset="UTF-8">
+<title>Train predictor statistics </title>
+<style>
+h1 {text-align: center;}
+
+form { 
+margin: 0 auto; 
+width:1200px;
+}
+div {
+border-radius: 50px;
+background-color: rgb(157, 206, 194);;
+padding: 5px;
+margin: 0 auto; 
+width:700px;
+}
+
+.button {
+border: none;
+color: rgb(0, 2, 1);
+padding: 16px 32px;
+text-align: center;
+text-decoration: none;
+display: inline-block;
+font-size: 14px;
+font: arial;
+margin: 6px 3px;
+transition-duration: 0.4s;
+cursor: pointer;
+border: 3px solid #818685;
+background-color: #bbbebd;
+border-radius: 15px
+}
+
+.button:hover {
+background-color: #549e7f;
+color: white;
+}
+
+.imgContainer{
+    float:left;
+}
+
+</style>
+</head>
+<body style="background: rgb(157, 206, 194);">
+<h1>Train predictor statistics </h1>
+<div>
+<div class="imgContainer">
+  <img src="/static/heatmap.png" alt="heatmap" width="550" height="380">
+</div>
+  <div class="imgContainer">
+  <img src="/static/mapcelcius.png" alt="heatmap" width="550" height="380">
+</div>
+<br>
+<form action="{{ url_for('home')}}">
+  <button type="submit" class="button">Back to prediction</button>
+</form>
+<br>
+</div>
+</body>
+</html>
